third_party/blink/web_tests/external/wpt/editing/run/insertunorderedlist_1-1000-expected.txt

This is a testharness.js-based test.
Found 242 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["insertunorderedlist",""]] "foo[]bar" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "foo[]bar" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "foo[bar]baz" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "foo[bar]baz" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "foo<br>[bar]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<ul><li>bar</li></ul>" but got "foo<br><ul><li>bar</li></ul>"
[FAIL] [["insertunorderedlist",""]] "foo<br>[bar]" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "foo<br>[bar]" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "f[oo<br>b]ar<br>baz" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "f[oo<br>b]ar<br>baz" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<p>[foo]<br>bar</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li></ul><p>bar</p>" but got "<p><ul><li>foo</li></ul>bar</p>"
[FAIL] [["insertunorderedlist",""]] "<p>[foo]<br>bar</p>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<p>[foo]<br>bar</p>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "[foo<ol><li>bar]</ol>baz" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "[foo<ol><li>bar]</ol>baz" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "foo<ol><li>[bar</ol>baz]" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "foo<ol><li>[bar</ol>baz]" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "[foo<ul><li>bar]</ul>baz" queryCommandIndeterm("insertunorderedlist") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["insertunorderedlist",""]] "[foo<ul><li>bar]</ul>baz" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "[foo<ul><li>bar]</ul>baz" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "foo<ul><li>[bar</ul>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<ul><li>bar</li><li>baz</li></ul>" but got "foo<br>bar<br><ul><li>baz</li></ul>"
[FAIL] [["insertunorderedlist",""]] "foo<ul><li>[bar</ul>baz]" queryCommandIndeterm("insertunorderedlist") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["insertunorderedlist",""]] "foo<ul><li>[bar</ul>baz]" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "foo<ul><li>[bar</ul>baz]" queryCommandState("insertunorderedlist") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["insertunorderedlist",""]] "foo<ul><li>[bar</ul>baz]" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "foo<ul><li>[bar</ul><ol><li>baz]</ol>quz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<ul><li>bar</li><li>baz</li></ul>quz" but got "foo<br>bar<br><ul><li>baz</li></ul>quz"
[FAIL] [["insertunorderedlist",""]] "foo<ul><li>[bar</ul><ol><li>baz]</ol>quz" queryCommandIndeterm("insertunorderedlist") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["insertunorderedlist",""]] "foo<ul><li>[bar</ul><ol><li>baz]</ol>quz" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "foo<ul><li>[bar</ul><ol><li>baz]</ol>quz" queryCommandState("insertunorderedlist") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["insertunorderedlist",""]] "foo<ul><li>[bar</ul><ol><li>baz]</ol>quz" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "foo<ol><li>[bar</ol><ul><li>baz]</ul>quz" queryCommandIndeterm("insertunorderedlist") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["insertunorderedlist",""]] "foo<ol><li>[bar</ol><ul><li>baz]</ul>quz" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "foo<ol><li>[bar</ol><ul><li>baz]</ul>quz" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<table><tbody><tr><td>foo<td>b[a]r<td>baz</table>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<table><tbody><tr><td>foo<td>b[a]r<td>baz</table>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<table><tbody><tr><td>fo[o<td>b]ar<td>baz</table>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<table><tbody><tr><td>fo[o<td>b]ar<td>baz</table>" queryCommandState("insertunorderedlist") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["insertunorderedlist",""]] "<table><tbody><tr><td>fo[o<td>b]ar<td>baz</table>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<p>foo<p>[bar]<p>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><ul><li>bar</li></ul><p>baz</p>" but got "<p>foo</p><p><ul><li>bar</li></ul></p><p>baz</p>"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<p>foo<p>[bar]<p>baz" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<p>foo<p>[bar]<p>baz" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<p>foo<p>[bar]<p>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><ul><li>bar</li></ul><p>baz</p>" but got "<p>foo</p><p><ul><li>bar</li></ul></p><p>baz</p>"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<p>foo<p>[bar]<p>baz" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<p>foo<p>[bar]<p>baz" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<p>foo<blockquote>[bar]</blockquote><p>baz" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<p>foo<blockquote>[bar]</blockquote><p>baz" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<dl><dt>foo<dd>[bar]<dt>baz<dd>quz</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt>foo</dt><dd><ul><li>bar</li></ul></dd><dt>baz</dt><dd>quz</dd></dl>" but got "<dl><dt>foo</dt>bar<br><dt>baz</dt><dd>quz</dd></dl>"
[FAIL] [["insertunorderedlist",""]] "<dl><dt>foo<dd>[bar]<dt>baz<dd>quz</dl>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<dl><dt>foo<dd>[bar]<dt>baz<dd>quz</dl>" queryCommandState("insertunorderedlist") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["insertunorderedlist",""]] "<dl><dt>foo<dd>[bar]<dt>baz<dd>quz</dl>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<dl><dt>foo<dd>bar<dt>[baz]<dd>quz</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt>foo</dt><dd>bar</dd><dt><ul><li>baz</li></ul></dt><dd>quz</dd></dl>" but got "<dl><dt>foo</dt><dd>bar</dd>baz<br><dd>quz</dd></dl>"
[FAIL] [["insertunorderedlist",""]] "<dl><dt>foo<dd>bar<dt>[baz]<dd>quz</dl>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<dl><dt>foo<dd>bar<dt>[baz]<dd>quz</dl>" queryCommandState("insertunorderedlist") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["insertunorderedlist",""]] "<dl><dt>foo<dd>bar<dt>[baz]<dd>quz</dl>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<p>[foo<p>bar]<p>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li><li>bar</li></ul><p>baz</p>" but got "<p><ul><li>foo</li><li>bar</li></ul></p><p>baz</p>"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<p>[foo<p>bar]<p>baz" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<p>[foo<p>bar]<p>baz" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<p>[foo<p>bar]<p>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li><li>bar</li></ul><p>baz</p>" but got "<p><ul><li>foo</li><li>bar</li></ul></p><p>baz</p>"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<p>[foo<p>bar]<p>baz" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<p>[foo<p>bar]<p>baz" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<p>[foo<blockquote>bar]</blockquote><p>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li><ul><li>bar</li></ul></ul><p>baz</p>" but got "<p><ul><li>foo</li><li>bar</li></ul></p><p>baz</p>"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<p>[foo<blockquote>bar]</blockquote><p>baz" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<p>[foo<blockquote>bar]</blockquote><p>baz" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<p>[foo<blockquote>bar]</blockquote><p>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li><ul><li>bar</li></ul></ul><p>baz</p>" but got "<p><ul><li>foo</li><li>bar</li></ul></p><p>baz</p>"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<p>[foo<blockquote>bar]</blockquote><p>baz" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<p>[foo<blockquote>bar]</blockquote><p>baz" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<dl><dt>[foo<dd>bar]<dt>baz<dd>quz</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt><ul><li>foo</li></ul></dt><dd><ul><li>bar</li></ul></dd><dt>baz</dt><dd>quz</dd></dl>" but got "<dl><ul><dt>foo</dt><dd>bar</dd></ul><dt>baz</dt><dd>quz</dd></dl>"
[FAIL] [["insertunorderedlist",""]] "<dl><dt>[foo<dd>bar]<dt>baz<dd>quz</dl>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<dl><dt>[foo<dd>bar]<dt>baz<dd>quz</dl>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<dl><dt>foo<dd>[bar<dt>baz]<dd>quz</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt>foo</dt><dd><ul><li>bar</li></ul></dd><dt><ul><li>baz</li></ul></dt><dd>quz</dd></dl>" but got "<dl><dt>foo</dt><ul><dd>bar</dd><dt>baz</dt></ul><dd>quz</dd></dl>"
[FAIL] [["insertunorderedlist",""]] "<dl><dt>foo<dd>[bar<dt>baz]<dd>quz</dl>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<dl><dt>foo<dd>[bar<dt>baz]<dd>quz</dl>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<p>[foo<blockquote><p>bar]<p>baz</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li><ul><li>bar</li></ul></ul><blockquote><p>baz</p></blockquote>" but got "<p><ul><li>foo</li><li>bar</li></ul></p><blockquote><p>baz</p></blockquote>"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<p>[foo<blockquote><p>bar]<p>baz</blockquote>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<p>[foo<blockquote><p>bar]<p>baz</blockquote>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<p>[foo<blockquote><p>bar]<p>baz</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li><ul><li>bar</li></ul></ul><blockquote><p>baz</p></blockquote>" but got "<p><ul><li>foo</li><li>bar</li></ul></p><blockquote><p>baz</p></blockquote>"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<p>[foo<blockquote><p>bar]<p>baz</blockquote>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<p>[foo<blockquote><p>bar]<p>baz</blockquote>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo<li>[bar]<li>baz</ol>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo<li>[bar]<li>baz</ol>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo</ol>[bar]" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo</ol>[bar]" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "[foo]<ol><li>bar</ol>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "[foo]<ol><li>bar</ol>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo</ol>[bar]<ol><li>baz</ol>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo</ol>[bar]<ol><li>baz</ol>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ol><ol><li>[foo]</ol></ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><ul><li>foo</li></ul></ul>" but got "<ol><ul><li>foo</li></ul></ol>"
[FAIL] [["insertunorderedlist",""]] "<ol><ol><li>[foo]</ol></ol>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<ol><ol><li>[foo]</ol></ol>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ol><li>[foo]<br>bar<li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo<br>bar</li></ul><ol><li>baz</li></ol>" but got "<ul><li>foo</li></ul>bar<br><ol><li>baz</li></ol>"
[FAIL] [["insertunorderedlist",""]] "<ol><li>[foo]<br>bar<li>baz</ol>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<ol><li>[foo]<br>bar<li>baz</ol>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo<br>[bar]<li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo<br>bar</li></ul><ol><li>baz</li></ol>" but got "foo<br><ul><li>bar</li></ul><ol><li>baz</li></ol>"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo<br>[bar]<li>baz</ol>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo<br>[bar]<li>baz</ol>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ol><li><div>[foo]</div>bar<li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li><div>foo</div>bar</li></ul><ol><li>baz</li></ol>" but got "<ul><li><div style=\\"display:inline !important\\">foo</div>bar</li></ul><ol><li>baz</li></ol>"
[FAIL] [["insertunorderedlist",""]] "<ol><li><div>[foo]</div>bar<li>baz</ol>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<ol><li><div>[foo]</div>bar<li>baz</ol>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo<ol><li>[bar]<li>baz</ol><li>quz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ul><li>bar</li></ul><ol><li>baz</li></ol><li>quz</li></ol>" but got "<ol><li>foo<ul><li>bar</li></ul><ol><li>baz</li></ol></li><li>quz</li></ol>"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo<ol><li>[bar]<li>baz</ol><li>quz</ol>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo<ol><li>[bar]<li>baz</ol><li>quz</ol>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo<ol><li>bar<li>[baz]</ol><li>quz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar</li></ol><ul><li>baz</li></ul><li>quz</li></ol>" but got "<ol><li>foo<ol><li>bar</li></ol><ul><li>baz</li></ul></li><li>quz</li></ol>"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo<ol><li>bar<li>[baz]</ol><li>quz</ol>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo<ol><li>bar<li>[baz]</ol><li>quz</ol>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo</li><ol><li>[bar]<li>baz</ol><li>quz</ol>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo</li><ol><li>[bar]<li>baz</ol><li>quz</ol>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo</li><ol><li>bar<li>[baz]</ol><li>quz</ol>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo</li><ol><li>bar<li>[baz]</ol><li>quz</ol>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ol><li>[foo]<ol><li>bar</ol><li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li></ul><ol><ol><li>bar</li></ol><li>baz</li></ol>" but got "<ul><li>foo</li></ul><ol><li>bar</li></ol><ol><li>baz</li></ol>"
[FAIL] [["insertunorderedlist",""]] "<ol><li>[foo]<ol><li>bar</ol><li>baz</ol>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<ol><li>[foo]<ol><li>bar</ol><li>baz</ol>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ol><li>[foo]</li><ol><li>bar</ol><li>baz</ol>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<ol><li>[foo]</li><ol><li>bar</ol><li>baz</ol>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo<li>[bar]<ol><li>baz</ol><li>quz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li></ol><ul><li>bar</li></ul><ol><ol><li>baz</li></ol><li>quz</li></ol>" but got "<ol><li>foo</li></ol><ul><li>bar</li></ul><ol><li>baz</li></ol><ol><li>quz</li></ol>"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo<li>[bar]<ol><li>baz</ol><li>quz</ol>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo<li>[bar]<ol><li>baz</ol><li>quz</ol>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo<li>[bar]</li><ol><li>baz</ol><li>quz</ol>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo<li>[bar]</li><ol><li>baz</ol><li>quz</ol>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo<ol><li>bar<li>baz</ol><li>[quz]</ol>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo<ol><li>bar<li>baz</ol><li>[quz]</ol>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo</li><ol><li>bar<li>baz</ol><li>[quz]</ol>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo</li><ol><li>bar<li>baz</ol><li>[quz]</ol>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo<li>[bar<li>baz]</ol>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo<li>[bar<li>baz]</ol>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ol><li>[foo<ol><li>bar]</ol><li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li><ul><li>bar</li></ul></ul><ol><li>baz</li></ol>" but got "<ul><li>foo</li><li>bar</li></ul><ol><li>baz</li></ol>"
[FAIL] [["insertunorderedlist",""]] "<ol><li>[foo<ol><li>bar]</ol><li>baz</ol>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<ol><li>[foo<ol><li>bar]</ol><li>baz</ol>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo<ol><li>b[ar</ol><li>b]az</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li></ol><ul><ul><li>bar</li></ul><li>baz</li></ul>" but got "<ol><li>foo<ul><li>bar</li></ul></li></ol><ul><li>baz</li></ul>"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo<ol><li>b[ar</ol><li>b]az</ol>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo<ol><li>b[ar</ol><li>b]az</ol>" queryCommandState("insertunorderedlist") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo<ol><li>b[ar</ol><li>b]az</ol>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<ol><li>[foo<ol><li>bar</ol><li>baz]</ol><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li><ul><li>bar</li></ul><li>baz</li></ul><p>extra</p>" but got "<ul><li>foo<ul><li>bar</li></ul></li><li>baz</li></ul><p>extra</p>"
[FAIL] [["insertunorderedlist",""]] "<ol><li>[foo<ol><li>bar</ol><li>baz]</ol><p>extra" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<ol><li>[foo<ol><li>bar</ol><li>baz]</ol><p>extra" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ol><li>[foo]<ol><li>bar</ol>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li></ul><ol><ol><li>bar</li></ol><li>baz</li></ol>" but got "<ul><li>foo</li></ul><ol><li>bar</li></ol>baz"
[FAIL] [["insertunorderedlist",""]] "<ol><li>[foo]<ol><li>bar</ol>baz</ol>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<ol><li>[foo]<ol><li>bar</ol>baz</ol>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo<ol><li>[bar]</ol>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ul><li>bar</li></ul><li>baz</li></ol>" but got "<ol><li>foo<ul><li>bar</li></ul>baz</li></ol>"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo<ol><li>[bar]</ol>baz</ol>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo<ol><li>[bar]</ol>baz</ol>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo<ol><li>bar</ol>[baz]</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar</li></ol></ol><ul><li>baz</li></ul>" but got "foo<ol><li>bar</li></ol><ul><li>baz</li></ul>"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo<ol><li>bar</ol>[baz]</ol>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo<ol><li>bar</ol>[baz]</ol>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ol><li>[foo<ol><li>bar]</ol>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li><ul><li>bar</li></ul></ul><ol><li>baz</li></ol>" but got "<ul><li>foo</li><li>bar</li></ul>baz"
[FAIL] [["insertunorderedlist",""]] "<ol><li>[foo<ol><li>bar]</ol>baz</ol>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<ol><li>[foo<ol><li>bar]</ol>baz</ol>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>foo<li>[bar]<li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li></ul><div>bar</div><ul><li>baz</li></ul>" but got "<ul><li>foo</li></ul>bar<br><ul><li>baz</li></ul>"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>foo<li>[bar]<li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li></ul><p>bar</p><ul><li>baz</li></ul>" but got "<ul><li>foo</li></ul>bar<br><ul><li>baz</li></ul>"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<ul><li>foo</ul>[bar]" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<ul><li>foo</ul>[bar]" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "[foo]<ul><li>bar</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "[foo]<ul><li>bar</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ul><li>foo</ul>[bar]<ul><li>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<ul><li>foo</ul>[bar]<ul><li>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ul><ul><li>[foo]</ul></ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ul><ul><li>[foo]</ul></ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>[foo]<br>bar<li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo<br>bar</div><ul><li>baz</li></ul>" but got "foo<br>bar<br><ul><li>baz</li></ul>"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>[foo]<br>bar<li>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>[foo]<br>bar<li>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>[foo]<br>bar<li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<br>bar</p><ul><li>baz</li></ul>" but got "foo<br>bar<br><ul><li>baz</li></ul>"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>[foo]<br>bar<li>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>[foo]<br>bar<li>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>foo<br>[bar]<li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo<br>bar</div><ul><li>baz</li></ul>" but got "foo<br>bar<br><ul><li>baz</li></ul>"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>foo<br>[bar]<li>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>foo<br>[bar]<li>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>foo<br>[bar]<li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<br>bar</p><ul><li>baz</li></ul>" but got "foo<br>bar<br><ul><li>baz</li></ul>"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>foo<br>[bar]<li>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>foo<br>[bar]<li>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li><div>[foo]</div>bar<li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo</div><div>bar</div><ul><li>baz</li></ul>" but got "<div style=\\"display:inline !important\\">foo</div>bar<br><ul><li>baz</li></ul>"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li><div>[foo]</div>bar<li>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li><div>[foo]</div>bar<li>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li><div>[foo]</div>bar<li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo</div><p>bar</p><ul><li>baz</li></ul>" but got "<div style=\\"display:inline !important\\">foo</div>bar<br><ul><li>baz</li></ul>"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li><div>[foo]</div>bar<li>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li><div>[foo]</div>bar<li>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<ul><li>foo<ul><li>[bar]<li>baz</ul><li>quz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li><li>bar</li><ul><li>baz</li></ul><li>quz</li></ul>" but got "<ul><li>foo<li>bar</li><ul><li>baz</li></ul></li><li>quz</li></ul>"
[FAIL] [["insertunorderedlist",""]] "<ul><li>foo<ul><li>[bar]<li>baz</ul><li>quz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ul><li>foo<ul><li>[bar]<li>baz</ul><li>quz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ul><li>foo<ul><li>bar<li>[baz]</ul><li>quz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li><ul><li>bar</li></ul><li>baz</li><li>quz</li></ul>" but got "<ul><li>foo<ul><li>bar</li></ul><li>baz</li></li><li>quz</li></ul>"
[FAIL] [["insertunorderedlist",""]] "<ul><li>foo<ul><li>bar<li>[baz]</ul><li>quz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ul><li>foo<ul><li>bar<li>[baz]</ul><li>quz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ul><li>foo</li><ul><li>[bar]<li>baz</ul><li>quz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ul><li>foo</li><ul><li>[bar]<li>baz</ul><li>quz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ul><li>foo</li><ul><li>bar<li>[baz]</ul><li>quz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ul><li>foo</li><ul><li>bar<li>[baz]</ul><li>quz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>[foo]<ul><li>bar</ul><li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo</div><ul><ul><li>bar</li></ul><li>baz</li></ul>" but got "foo<ul><li>bar</li></ul><ul><li>baz</li></ul>"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>[foo]<ul><li>bar</ul><li>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>[foo]<ul><li>bar</ul><li>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>[foo]<ul><li>bar</ul><li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><ul><ul><li>bar</li></ul><li>baz</li></ul>" but got "foo<ul><li>bar</li></ul><ul><li>baz</li></ul>"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>[foo]<ul><li>bar</ul><li>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>[foo]<ul><li>bar</ul><li>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>[foo]</li><ul><li>bar</ul><li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo</div><ul><ul><li>bar</li></ul><li>baz</li></ul>" but got "foo<br><ul><ul><li>bar</li></ul><li>baz</li></ul>"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>[foo]</li><ul><li>bar</ul><li>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>[foo]</li><ul><li>bar</ul><li>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>[foo]</li><ul><li>bar</ul><li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><ul><ul><li>bar</li></ul><li>baz</li></ul>" but got "foo<br><ul><ul><li>bar</li></ul><li>baz</li></ul>"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>[foo]</li><ul><li>bar</ul><li>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>[foo]</li><ul><li>bar</ul><li>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>foo<li>[bar]<ul><li>baz</ul><li>quz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li></ul><div>bar</div><ul><ul><li>baz</li></ul><li>quz</li></ul>" but got "<ul><li>foo</li></ul>bar<ul><li>baz</li></ul><ul><li>quz</li></ul>"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>foo<li>[bar]<ul><li>baz</ul><li>quz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>foo<li>[bar]<ul><li>baz</ul><li>quz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>foo<li>[bar]<ul><li>baz</ul><li>quz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li></ul><p>bar</p><ul><ul><li>baz</li></ul><li>quz</li></ul>" but got "<ul><li>foo</li></ul>bar<ul><li>baz</li></ul><ul><li>quz</li></ul>"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>foo<li>[bar]<ul><li>baz</ul><li>quz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>foo<li>[bar]<ul><li>baz</ul><li>quz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>foo<li>[bar]</li><ul><li>baz</ul><li>quz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li></ul><div>bar</div><ul><ul><li>baz</li></ul><li>quz</li></ul>" but got "<ul><li>foo</li></ul>bar<br><ul><ul><li>baz</li></ul><li>quz</li></ul>"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>foo<li>[bar]</li><ul><li>baz</ul><li>quz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>foo<li>[bar]</li><ul><li>baz</ul><li>quz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>foo<li>[bar]</li><ul><li>baz</ul><li>quz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li></ul><p>bar</p><ul><ul><li>baz</li></ul><li>quz</li></ul>" but got "<ul><li>foo</li></ul>bar<br><ul><ul><li>baz</li></ul><li>quz</li></ul>"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>foo<li>[bar]</li><ul><li>baz</ul><li>quz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>foo<li>[bar]</li><ul><li>baz</ul><li>quz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>foo<ul><li>bar<li>baz</ul><li>[quz]</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo<ul><li>bar</li><li>baz</li></ul></li></ul><div>quz</div>" but got "<ul><li>foo<ul><li>bar</li><li>baz</li></ul></li></ul>quz"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>foo<ul><li>bar<li>baz</ul><li>[quz]</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>foo<ul><li>bar<li>baz</ul><li>[quz]</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>foo<ul><li>bar<li>baz</ul><li>[quz]</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo<ul><li>bar</li><li>baz</li></ul></li></ul><p>quz</p>" but got "<ul><li>foo<ul><li>bar</li><li>baz</li></ul></li></ul>quz"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>foo<ul><li>bar<li>baz</ul><li>[quz]</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>foo<ul><li>bar<li>baz</ul><li>[quz]</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>foo</li><ul><li>bar<li>baz</ul><li>[quz]</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li><ul><li>bar</li><li>baz</li></ul></ul><div>quz</div>" but got "<ul><li>foo</li><ul><li>bar</li><li>baz</li></ul></ul>quz"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>foo</li><ul><li>bar<li>baz</ul><li>[quz]</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>foo</li><ul><li>bar<li>baz</ul><li>[quz]</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>foo</li><ul><li>bar<li>baz</ul><li>[quz]</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li><ul><li>bar</li><li>baz</li></ul></ul><p>quz</p>" but got "<ul><li>foo</li><ul><li>bar</li><li>baz</li></ul></ul>quz"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>foo</li><ul><li>bar<li>baz</ul><li>[quz]</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>foo</li><ul><li>bar<li>baz</ul><li>[quz]</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>foo<li>[bar<li>baz]</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li></ul><div>bar</div><div>baz</div>" but got "<ul><li>foo</li></ul>bar<br>baz"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>foo<li>[bar<li>baz]</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>foo<li>[bar<li>baz]</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>foo<li>[bar<li>baz]</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li></ul><p>bar</p><p>baz</p>" but got "<ul><li>foo</li></ul>bar<br>baz"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>foo<li>[bar<li>baz]</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>foo<li>[bar<li>baz]</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>[foo<ul><li>bar]</ul><li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo</div><ul><li>bar</li><li>baz</li></ul>" but got "foo<br>bar<br><ul><li>baz</li></ul>"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>[foo<ul><li>bar]</ul><li>baz</ul>" queryCommandState("insertunorderedlist") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>[foo<ul><li>bar]</ul><li>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>[foo<ul><li>bar]</ul><li>baz</ul>" queryCommandIndeterm("insertunorderedlist") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>[foo<ul><li>bar]</ul><li>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>[foo<ul><li>bar]</ul><li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><ul><li>bar</li><li>baz</li></ul>" but got "foo<br>bar<br><ul><li>baz</li></ul>"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>[foo<ul><li>bar]</ul><li>baz</ul>" queryCommandState("insertunorderedlist") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>[foo<ul><li>bar]</ul><li>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>[foo<ul><li>bar]</ul><li>baz</ul>" queryCommandIndeterm("insertunorderedlist") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>[foo<ul><li>bar]</ul><li>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>foo<ul><li>b[ar</ul><li>b]az</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li><li>bar</li></ul><div>baz</div>" but got "<ul><li>foo<li>bar</li></li></ul>baz"
Harness: the test ran to completion.
